paragraph divisions and spacing
As you probably noticed, paragraph spacing and newlines don't work the same in LaTeX as they do in e.g. word. The correct style (inasmuch as any style is correct - at least the prevailing style) is to have each paragraph be one line, and new paragraphs are created by leaving an empty line.

As you noticed, a single line break is just completely ignored by LaTeX -- so since this is confusing, we should avoid having line breaks in the TeX code that don't actually do anything. (I have had at least one typo in a paper I wrote caused by ignoring this practice!)

The \\ to create a new line should be very sparingly used -- only as in your authors list, and in math mode.

To create more space, such as after the exercise statements, we can use the \vspace and \hspace commands to create vertical and horizontal spacing.

(I also fixed two very minor typos in this commit - sorry for the nonatomicity of the commit...)