fix spacing and linebreaks
Using \\ is almost always bad practice outside of an equation. Using line breaks in the TeX that do not change the appearance of the document is *always* bad practice, and has led to actual errors (albeit minor) for me in the past in the mathematical content of what I was writing.

Using \noindent should be done very sparingly - it looks noticeably odd to have different indentation than standard TeX, so it should be done for a reason.