language tweaks and TeX fixes
The main comment here is that *whenever you write a maths symbol, it must be in an equation* - TeX uses visibly different fonts for G and $G$, so whenever you are talking about a graph named G in TeX, you write its name as $G$, to get the right font for it. Not doing so looks unprofessional.